requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-google-genai
pypdf
python-dotenv
numpy
sentence-transformers
chromadb